[441] Move NoOp implementation of ICustomImagesService into its own class
Bug: https://github.com/eclipse-sirius/sirius-components/issues/441
