+++ backend/sirius-web-spring/src/test/java/org/eclipse/sirius/web/spring/controllers/ImagesControllerTestCases.java
@@ -24,7 +24,6 @@ import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 
 import org.eclipse.sirius.web.api.services.IImagePathService;
-import org.eclipse.sirius.web.services.api.images.ICustomImagesService;
 import org.junit.Test;
 import org.springframework.core.io.Resource;
 import org.springframework.http.HttpHeaders;
@@ -45,7 +44,7 @@ import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
 public class ImagesControllerTestCases {
     @Test
     public void testInvalidFolder() {
-        ImagesController imagesController = new ImagesController(new ArrayList<>(), new ICustomImagesService.NoOp(), new SimpleMeterRegistry());
+        ImagesController imagesController = new ImagesController(new ArrayList<>(), new NoOpCustomService(), new SimpleMeterRegistry());
         HttpServletRequest request = new MockHttpServletRequest(HttpMethod.GET.name(), "/api/images/invalidFolder/doesNotExist.png"); //$NON-NLS-1$
         ResponseEntity<Resource> responseEntity = imagesController.getImage(request);
         assertThat(responseEntity.getStatusCode()).isEqualTo(HttpStatus.NOT_FOUND);
@@ -59,7 +58,7 @@ public class ImagesControllerTestCases {
                 return List.of("/validFolder"); //$NON-NLS-1$
             }
         };
-        ImagesController imagesController = new ImagesController(List.of(imagePathService), new ICustomImagesService.NoOp(), new SimpleMeterRegistry());
+        ImagesController imagesController = new ImagesController(List.of(imagePathService), new NoOpCustomService(), new SimpleMeterRegistry());
         HttpServletRequest request = new MockHttpServletRequest(HttpMethod.GET.name(), "/api/images/validFolder/doesNotExist.png"); //$NON-NLS-1$
         ResponseEntity<Resource> responseEntity = imagesController.getImage(request);
         assertThat(responseEntity.getStatusCode()).isEqualTo(HttpStatus.NOT_FOUND);
@@ -73,7 +72,7 @@ public class ImagesControllerTestCases {
                 return List.of("/icons"); //$NON-NLS-1$
             }
         };
-        ImagesController imagesController = new ImagesController(List.of(imagePathService), new ICustomImagesService.NoOp(), new SimpleMeterRegistry());
+        ImagesController imagesController = new ImagesController(List.of(imagePathService), new NoOpCustomService(), new SimpleMeterRegistry());
         HttpServletRequest request = new MockHttpServletRequest(HttpMethod.GET.name(), "/api/images/icons/full/obj16/EClass.gif"); //$NON-NLS-1$
 
         // We need to replace the current class loader to trick Spring into thinking that the resource exists
